# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MADHack Agon API",
    "description": "MADHack Initial Round API Documentation",
    "contact": {
      "email": "ucscieee@gmail.com"
    },
    "version": "1.2.0"
  },
  "servers": [
    {
      "url": "https://agon.madhack.ucscieee.com/api",
      "description": "Agon Server"
    },
    {
      "url": "https://backup.agon.madhack.ucscieee.com/api",
      "description": "Agon Backup Server"
    },
    {
      "url": "http://localhost:3001/api",
      "description": "Local Server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    },
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/todo": {
      "get": {
        "tags": [
          "todo"
        ],
        "summary": "retrieves todo list",
        "description": "Gets todo list of logged in user",
        "operationId": "getTodoList",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListOfTodos"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_401_object"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_500_object"
                },
                "examples": {
                  "internal-error": {
                    "value": {
                      "error": "Internal Server Error"
                    }
                  },
                  "get-error": {
                    "value": {
                      "error": "Error getting todo list"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "todo"
        ],
        "summary": "creates a todo",
        "description": "Creates a todo",
        "operationId": "createTodo",
        "requestBody": {
          "description": "Todo details",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Todo"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_201_object"
                },
                "example": {
                  "message": "Todo created successfully"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_400_object"
                },
                "example": {
                  "error": "fieldName is required"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_401_object"
                },
                "examples": {
                  "unauthorized": {
                    "value": {
                      "error": "No auth token provided"
                    }
                  },
                  "invalid-token": {
                    "value": {
                      "error": "API access denied! No team access token provided"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_500_object"
                },
                "examples": {
                  "internal-error": {
                    "value": {
                      "error": "Internal Server Error"
                    }
                  },
                  "create-error": {
                    "value": {
                      "error": "Error creating todo"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/todo/{todoId}": {
      "parameters": [
        {
          "name": "todoId",
          "in": "path",
          "description": "ID of todo to return",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "tags": [
          "todo"
        ],
        "summary": "retrieves todo list",
        "description": "Gets todo list of logged in user",
        "operationId": "getTodo",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "BearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "todo"
        ],
        "summary": "updates a todo",
        "description": "Updates a todo",
        "operationId": "updateTodo",
        "requestBody": {
          "description": "Todo details",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoWithoutId"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_200_object"
                },
                "example": {
                  "message": "Todo updated successfully"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_400_object"
                },
                "example": {
                  "error": "fieldName is required"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_401_object"
                },
                "examples": {
                  "unauthorized": {
                    "value": {
                      "error": "No auth token provided"
                    }
                  },
                  "invalid-token": {
                    "value": {
                      "error": "API access denied! No team access token provided"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_500_object"
                },
                "examples": {
                  "internal-error": {
                    "value": {
                      "error": "Internal Server Error"
                    }
                  },
                  "update-error": {
                    "value": {
                      "error": "Error updating todo"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "todo"
        ],
        "summary": "deletes a todo",
        "description": "Deletes a todo",
        "operationId": "deleteTodo",
        "requestBody": {
          "description": "Todo details",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Todo"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_200_object"
                },
                "example": {
                  "message": "Todo deleted successfully"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_401_object"
                },
                "examples": {
                  "unauthorized": {
                    "value": {
                      "error": "No auth token provided"
                    }
                  },
                  "invalid-token": {
                    "value": {
                      "error": "API access denied! No team access token provided"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_500_object"
                },
                "examples": {
                  "internal-error": {
                    "value": {
                      "error": "Internal Server Error"
                    }
                  },
                  "delete-error": {
                    "value": {
                      "error": "Error deleting todo"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/todo/{todoId}/status": {
      "parameters": [
        {
          "name": "todoId",
          "in": "path",
          "description": "ID of todo to return",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "put": {
        "tags": [
          "todo"
        ],
        "summary": "updates a todo status",
        "description": "Updates the completion status for a specific todo",
        "operationId": "updateTodoStatus",
        "requestBody": {
          "description": "Todo status",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoStatus"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_200_object"
                },
                "example": {
                  "message": "Todo status updated successfully"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_400_object"
                },
                "example": {
                  "error": "isCompleted is required"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_401_object"
                },
                "examples": {
                  "unauthorized": {
                    "value": {
                      "error": "No auth token provided"
                    }
                  },
                  "invalid-token": {
                    "value": {
                      "error": "API access denied! No team access token provided"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_500_object"
                },
                "examples": {
                  "internal-error": {
                    "value": {
                      "error": "Internal Server Error"
                    }
                  },
                  "update-error": {
                    "value": {
                      "error": "Error updating todo"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ListOfTodos": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Todo"
        }
      },
      "Todo": {
        "required": [
          "todoId",
          "categoryId",
          "title",
          "notes",
          "createdDt",
          "dueDt",
          "isCompleted",
          "isReminderEnabled",
          "lastModifiedDt"
        ],
        "type": "object",
        "properties": {
          "todoId": {
            "type": "string",
            "format": "uuid",
            "example": "d290f1ee-6c54-4b01-90e6-d701748f0851"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid",
            "example": "064596ab-85ce-4107-b390-2e3989f7ef40"
          },
          "title": {
            "type": "string",
            "example": "Buy Milk"
          },
          "notes": {
            "type": "string",
            "example": "Buy 2L of milk"
          },
          "createdDt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-02-25T10:00:00Z"
          },
          "dueDt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-02-25T10:00:00Z"
          },
          "isCompleted": {
            "type": "boolean",
            "example": false
          },
          "isReminderEnabled": {
            "type": "boolean",
            "example": false
          },
          "lastModifiedDt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-02-25T10:00:00Z"
          }
        }
      },
      "TodoWithoutId": {
        "required": [
          "categoryId",
          "title",
          "notes",
          "createdDt",
          "dueDt",
          "isCompleted",
          "isReminderEnabled",
          "lastModifiedDt"
        ],
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string",
            "format": "uuid",
            "example": "064596ab-85ce-4107-b390-2e3989f7ef40"
          },
          "title": {
            "type": "string",
            "example": "Buy Milk"
          },
          "notes": {
            "type": "string",
            "example": "Buy 2L of milk"
          },
          "createdDt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-02-25T10:00:00Z"
          },
          "dueDt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-02-25T10:00:00Z"
          },
          "isCompleted": {
            "type": "boolean",
            "example": false
          },
          "isReminderEnabled": {
            "type": "boolean",
            "example": false
          },
          "lastModifiedDt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-02-25T10:00:00Z"
          }
        }
      },
      "TodoStatus": {
        "required": [
          "isCompleted"
        ],
        "type": "object",
        "properties": {
          "isCompleted": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "response_200_object": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "response_201_object": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Object created successfully"
          }
        }
      },
      "response_401_object": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "API access denied! No team access token provided"
          }
        }
      },
      "response_400_object": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Bad Request"
          }
        }
      },
...
      "response_500_object": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Internal Server Error"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "name": "X-API-Key",
        "in": "header"
      }
    }
  }
}